requirements.txt
openai>=1.0.0
python-dotenv
tqdm
pydantic<2.0.0
numpy
pypinyin
